+++ src/elf/cache.rs
@@ -28,6 +28,17 @@ pub(crate) enum ElfBackend {
     Elf(Rc<ElfParser>), // ELF w/o DWARF
 }
 
+impl ElfBackend {
+    /// Retrieve the underlying [`ElfParser`].
+    pub(crate) fn parser(&self) -> &ElfParser {
+        match self {
+            #[cfg(feature = "dwarf")]
+            Self::Dwarf(resolver) => resolver.parser(),
+            Self::Elf(parser) => parser,
+        }
+    }
+}
+
 #[cfg(test)]
 #[cfg(feature = "dwarf")]
 impl ElfBackend {